Description:-
This C++ program on BANKING SYSTEM has an account class with data members like account number, name, deposit, withdraw amount and type of account. Customer data is stored in a binary file. A customer can deposit and withdraw the amount in his account. Users can create, modify and delete the account.
In this banking system project, we have not used graphics to keep the program simple.